requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Machine Learning
scikit-learn>=1.2.0

# Deep Learning (optional - uncomment if needed)
# torch>=2.0.0
# tensorflow>=2.12.0

# Jupyter
jupyterlab>=4.0.0
ipykernel>=6.22.0
ipywidgets>=8.0.0

# Data Validation
great-expectations>=0.17.0

# Code Quality
pytest>=7.3.0
pytest-cov>=4.0.0
flake8>=6.0.0
black>=23.3.0
isort>=5.12.0

# Automation
python-dotenv>=1.0.0
openai>=1.0.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.28.0
